Store all strategies in the Strategies.py file.
