+++ UnleashClient/strategies/Strategies.py
@@ -1,11 +1,53 @@
+# pylint: disable=dangerous-default-value
 from UnleashClient.constraints import Constraint
 
-# pylint: disable=dangerous-default-value
+
+class Strategy():
+    """
+    In general, default & custom classes should only need to override:
+    * __call__() - Implementation of the strategy.
+    * load_provisioning - Loads strategy provisioning
+    """
+    def __init__(self,
+                 parameters: dict = {}) -> None:
+        """
+        A generic strategy objects.
+
+        :param parameters: 'parameters' key from strategy section (...from feature section) of
+        /api/clients/features response
+        """
+        # Experiment information
+        self.parameters = parameters
+
+        self.parsed_provisioning = self.load_provisioning()
+
+    # pylint: disable=no-self-use
+    def load_provisioning(self) -> list:
+        """
+        Method to load data on object initialization, if desired.
+
+        This should parse the raw values in self.parameters into format Python can comprehend.
+        """
+        return []
+
+    def __eq__(self, other):
+        return self.parameters == other.parameters
+
+    def __call__(self, context: dict = None) -> bool:
+        """
+        Strategy implementation goes here.
+
+        :param context: Context information
+        :return:
+        """
+        return False
+
+
 class StrategyV2:
     """
     In general, default & custom classes should only need to override:
     * __init__() - Depending on the parameters your feature needs
-    * strategy() - Your feature provisioning
+    * apply_strategy() - Your feature provisioning
     """
     def __init__(self,
                  constraints: list = [],
@@ -32,7 +74,7 @@ class StrategyV2:
         flag_state = False
 
         if all([constraint(context) for constraint in self.parsed_constraints]):
-            flag_state = self.strategy(context)
+            flag_state = self.apply_strategy(context)
 
         return flag_state
 
@@ -58,7 +100,7 @@ class StrategyV2:
         """
         return []
 
-    def strategy(self, context: dict = None) -> bool:  #pylint: disable=W0613,R0201
+    def apply_strategy(self, context: dict = None) -> bool:  #pylint: disable=W0613,R0201
         """
         Strategy implementation goes here.
